+++ package.json
@@ -17,6 +17,7 @@
     "@mui/styled-engine-sc": "^6.0.0-alpha.10",
     "@types/multiparty": "^0.0.36",
     "@types/socket.io": "^3.0.2",
+    "@types/styled-components": "^5.1.34",
     "@vercel/postgres": "^0.5.1",
     "axios": "^1.6.3",
     "bcryptjs": "^2.4.3",
@@ -42,7 +43,7 @@
     "@types/bcryptjs": "^2.4.6",
     "@types/mime-types": "^2.1.4",
     "@types/node": "^20",
-    "@types/react": "^18",
+    "@types/react": "^18.2.48",
     "@types/react-dom": "^18",
     "autoprefixer": "^10.0.1",
     "eslint": "^8",
