+++ src/app/create/page.tsx
@@ -1,0 +1,18 @@
+"use client";
+import Layout from "@/components/Layout";
+import CreateProduct from "./Create";
+
+function NewProduct() {
+
+
+    return (
+        <Layout>
+            <h1>New Product</h1>
+            <CreateProduct />
+
+        </Layout>
+    );
+}
+
+
+export default NewProduct;
