+++ src/app/products/[...item]/page.tsx
@@ -1,0 +1,15 @@
+import Layout from "@/components/Layout";
+import { useRouter } from "next/dist/client/router";
+
+export default function ItemPage() {
+    const router = useRouter();
+    const { id } = router.query;
+    return (
+        <>
+            <Layout>
+                <h1>Item</h1>
+                {id}
+            </Layout>
+        </>
+    );
+}
